TT-355: Fix session stubs in feature helper
- The session() method defined in feature_helper was clashing with the
  built-in Rails session object (accessed as a hash in controller_helper)
  so we renamed it.
- ViewableIdentityProvider was being recreated in controller_helper
  unnecessarily and was missing the viewable?() method.
- Show LOA2 transactions that will require extra security on LOA1
  confirmation page.

Authors: @DataMinerUK @vixus0 @javindo
