+++ app/views/confirmation/confirmation_LOA1.html.erb
@@ -17,7 +17,7 @@
 
   <div class="panel panel-border-narrow">
     <ul class="list-bullet list">
-      <% loa1_transactions_list.each do |transaction| -%>
+      <% loa2_transactions_list.each do |transaction| -%>
           <li><%= transaction.name %></li>
       <% end %>
     </ul>
