+++ spec/controllers/confirmation_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
 require 'controller_helper'
 require 'spec_helper'
+require 'models/display/viewable_identity_provider'
 
 describe ConfirmationController do
   let(:identity_provider_display_decorator) { double(:IdentityProviderDisplayDecorator) }
@@ -14,8 +15,8 @@ describe ConfirmationController do
   subject { get :index, params: { locale: 'en' } }
 
   before(:each) do
-    create_session_cookie('selected_idp', 'entity_id' => entity_id, 'simple_id' => simple_id, 'levels_of_assurance' => levels_of_assurance)
-    create_session_cookie('transaction_simple_id', transaction_simple_id)
+    session[:selected_idp] = {'entity_id' => entity_id, 'simple_id' => simple_id, 'levels_of_assurance' => levels_of_assurance}
+    session[:transaction_simple_id] = transaction_simple_id
 
     stub_const('IDENTITY_PROVIDER_DISPLAY_DECORATOR', stub_identity_provider_display_decorator(identity_provider_display_decorator, simple_id, entity_id, levels_of_assurance))
     stub_const('RP_DISPLAY_REPOSITORY', stub_rp_display_repository(transaction_simple_id))
@@ -34,4 +35,31 @@ describe ConfirmationController do
     expect(subject).to render_template(:confirmation_LOA2)
     expect(subject).to_not render_template(:confirmation_LOA1)
   end
+end
+
+
+def stub_identity_provider_display_decorator(identity_provider_display_decorator, simple_id, entity_id, levels_of_assurance)
+  loa_identity_provider = IdentityProvider.new('simple_id' => simple_id,
+                                               'entity_id' => entity_id,
+                                               'levels_of_assurance' => levels_of_assurance)
+
+  viewable_identity_provider_stub = Display::ViewableIdentityProvider.new(loa_identity_provider, display_data, 'idp-logos/barclays.png', 'idp-logos-white/barclays.png')
+
+  expect(identity_provider_display_decorator).to receive(:decorate) { |identity_provider|
+    expect(identity_provider).to have_attributes(simple_id: simple_id,
+                                                 entity_id: entity_id,
+                                                 levels_of_assurance: levels_of_assurance)
+  }.and_return(viewable_identity_provider_stub)
+  expect(viewable_identity_provider_stub).to receive(:display_name).and_return('idp-display-name')
+
+  identity_provider_display_decorator
+end
+
+def stub_rp_display_repository(transaction_simple_data)
+  current_transaction = ""
+
+  def current_transaction.name
+    "Test-RP"
+  end
+  { transaction_simple_data => current_transaction }
 end
