+++ spec/features/redirects_with_see_other_spec.rb
@@ -2,7 +2,7 @@ require 'feature_helper'
 require 'api_test_helper'
 describe 'pages redirect with see other', type: :request do
   it 'sets a see other for redirects' do
-    stub_request(:post, api_uri('session')).to_return(body: session.to_json, status: 201)
+    stub_request(:post, api_uri('session')).to_return(body: stub_api_session.to_json, status: 201)
     stub_api_saml_endpoint
     post '/SAML2/SSO', params: { 'SAMLRequest' => 'my-saml-request', 'RelayState' => 'my-relay-state' }
     expect(response.status).to eql 303
